[DEV-3397] Refactoring
