+++ src/subdomains/core/buy-crypto/process/services/buy-crypto.service.ts
@@ -37,7 +37,7 @@ import { CheckoutTx } from 'src/subdomains/supporting/fiat-payin/entities/checko
 import { CheckoutTxService } from 'src/subdomains/supporting/fiat-payin/services/checkout-tx.service';
 import { CryptoInput } from 'src/subdomains/supporting/payin/entities/crypto-input.entity';
 import { PayInService } from 'src/subdomains/supporting/payin/services/payin.service';
-import { UpdateTransactionInternalDto } from 'src/subdomains/supporting/payment/dto/input/update-transaction.dto';
+import { UpdateTransactionInternalDto } from 'src/subdomains/supporting/payment/dto/input/update-transaction-internal.dto';
 import { TransactionRequest } from 'src/subdomains/supporting/payment/entities/transaction-request.entity';
 import { TransactionTypeInternal } from 'src/subdomains/supporting/payment/entities/transaction.entity';
 import { SpecialExternalAccountService } from 'src/subdomains/supporting/payment/services/special-external-account.service';
