+++ src/subdomains/supporting/payment/services/transaction.service.ts
@@ -3,7 +3,7 @@ import { Util } from 'src/shared/utils/util';
 import { UpdateTransactionDto } from 'src/subdomains/core/history/dto/update-transaction.dto';
 import { Between, FindOptionsRelations, IsNull, LessThanOrEqual, Not } from 'typeorm';
 import { CreateTransactionDto } from '../dto/input/create-transaction.dto';
-import { UpdateTransactionInternalDto } from '../dto/input/update-transaction.dto';
+import { UpdateTransactionInternalDto } from '../dto/input/update-transaction-internal.dto';
 import { Transaction } from '../entities/transaction.entity';
 import { TransactionRepository } from '../repositories/transaction.repository';
